requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
azure-core==1.34.0
azure-data-tables==12.7.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
coverage==7.9.1
cryptography==45.0.4
cycler==0.12.1
databricks-sdk==0.57.0
decorator==5.2.1
docker==7.1.0
executing==2.2.0
fastapi==0.115.12
flake8==7.2.0
Flask==3.1.1
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
JSON-log-formatter==1.1.1
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mlflow==3.1.0
mlflow-skinny==3.1.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
narwhals==1.43.0
numpy==2.3.0
nvidia-nccl-cu12==2.27.3
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.4.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
svgwrite==1.4.3
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
Tree==0.2.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.2
yarl==1.20.1
zipp==3.23.0